Hack to make `go build` work when using a weird directory structure.
`go build` by default compiles the main package into a binary named
the same as the current directory. Sometimes folks do things to
cause this directory to not be named "logstash-forwarder"
(for example, as reported in #388).

This change makes all `go build` documentation/invocations force the
output binary name to be "logstash-forwarder" regardless of the
parent directory name.